requirements/base.txt
# Bleeding edge Django
django==1.8.2

# Configuration
django-environ==0.3.0
django-secure==1.0.1
whitenoise==1.0.6


# Forms
django-braces==1.8.0
django-crispy-forms==1.4.0
django-floppyforms==1.4.1

# Models
django-model-utils==2.2

# Images
Pillow==2.8.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.20.0

# For the persistence stores
psycopg2==2.6

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.8.0

# Time zones support
pytz==2015.4

# Your custom requirements go here

bootstrap-admin==0.3.5
# django-admin-bootstrapped==2.3.5
# django-bootstrap3==5.1.1
django-localflavor==1.1
django-ajax-selects==1.3.6
django-ckeditor==4.4.7
django-imagekit==3.2.6
django-pagination==1.0.7
# django-taggit==0.13.0
# django-taggit-templatetags==0.2.5
# django-templatetag-sugar==1.0
# django-vanilla-views==1.0.3	
# docutils==0.12
django-avatar==2.0